+++ src/test/java/com/firebase/geofire/util/SimpleFuture.java
@@ -1,4 +1,4 @@
-package com.firebase.geofire.core;
+package com.firebase.geofire.util;
 
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
